requirements.txt
Django
requests
django-extensions
factory_boy
